feat: add register component

+++ src/app/features/auth/register/register.ts
@@ -4,25 +4,29 @@ import { AuthService } from '../../../core/services/auth';
 import { Router, RouterLink } from '@angular/router';
 
 @Component({
-  selector: 'app-login',
+  selector: 'app-register',
   standalone: true,
   imports: [FormsModule, RouterLink],
-  templateUrl: './login.html',
-  styleUrl: './login.scss',
+  templateUrl: './register.html',
+  styleUrl: './register.scss',
 })
-export class LoginComponent {
-  credentials = { email: '', password: '' };
+export class RegisterComponent {
+  userData = { name: '', email: '', password: '', password_confirmation: '' };
   errorMessage: string | null = null;
 
   constructor(private authService: AuthService, private router: Router) {}
 
-  onLogin() {
+  onRegister() {
     this.errorMessage = null;
 
-    this.authService.login(this.credentials).subscribe({
+    if (this.userData.password !== this.userData.password_confirmation) {
+      this.errorMessage = 'Password and password confirmation are not identical!';
+      return;
+    }
+
+    this.authService.register(this.userData).subscribe({
       next: (res) => {
         localStorage.setItem('access_token', res.access_token);
-        console.log('Successfully logged in!');
         this.router.navigate(['/dashboard']);
       },
       error: (err) => {
